+++ Blaster Master/Eyeball.cpp
@@ -1,5 +1,5 @@
 #include "Eyeball.h"
-void Eyeball::Update(DWORD dt)
+void Eyeball::Update(float dt)
 {
 
 }
