+++ Blaster Master/Eyeball.h
@@ -4,7 +4,7 @@ class Eyeball :
     public DynamicObject
 {
 public:
-	void Update(DWORD dt);
+	void Update(float dt);
 	void Render();
 };
