+++ Blaster Master/Worm.h
@@ -3,7 +3,7 @@
 class Worm : public DynamicObject
 {
 public:
-	void Update(DWORD dt);
+	void Update(float dt);
 	void Render();
 };
